refactor(harness): split result assembly and decompose measurement
Complete the Wave 2 measurement core (S9).

1. Split the 357-line _build_result on measurement source. The offline
   producer (build_offline_metrics) reads the InferenceOutput and packs
   everything source-derived into a SourceMetrics; the mode-agnostic
   assembler (assemble_experiment_result) takes SourceMetrics + baseline +
   identity + thermal and builds the ExperimentResult. The assembler
   signature carries no InferenceOutput - server mode adds a sibling
   producer without touching it.

2. Decompose the god-module (measurement.py 1431L) along phase seams into
   lifecycle (baseline/load/prompts/warmup), window (bracket orchestration +
   flops), result_assembly (item 1 + warnings), and persistence (sidecar
   writes). MeasurementHarness stays the facade; harness/__init__ re-exports
   are intact. Tests patch at the new use sites.

3. Resolve the batch-size contradiction to one SSOT. ssot.EngineDescriptor
   gains a BatchSizeModel encoding both facts: the declared capacity bound
   (VRAM worst-case) and the static configured batch (result field), which
   differ only for vLLM. Both consumers (cli/_vram, result assembly) read it
   via ExperimentConfig.capacity_batch_size() / static_batch_size().

No behavior or results change. Full unit + integration suites green; the
replay fixture and window-ordering tests pass unmodified except patch targets.
